Renamed post_objects -> post_objects_container.
